+++ oshi-json/src/main/java/oshi/json/hardware/CentralProcessor.java
@@ -18,6 +18,7 @@
  */
 package oshi.json.hardware;
 
+import oshi.hardware.ComputerSystem;
 import oshi.json.json.OshiJsonObject;
 
 /**
@@ -282,13 +283,19 @@ public interface CentralProcessor extends OshiJsonObject {
     long getSystemUptime();
 
     /**
-     * Get the System/CPU Serial Number, if available. On Linux, this requires
-     * either root permissions, or installation of the (deprecated) HAL library
-     * (lshal command).
+     * Get the System/CPU Serial Number, if available.
+     *
+     * On Linux and FreeBSD, this requires either root permissions, installation
+     * of the (deprecated) HAL library (lshal command). Linux also attempts to
+     * read the dmi/id serial number files in sysfs, which are read-only root by
+     * default but may have permissions altered by the user.
      *
      * @return the System/CPU Serial Number, if available, otherwise returns
      *         "unknown"
+     * 
+     * @deprecated use {@link ComputerSystem#getSerialNumber()} instead.
      */
+    @Deprecated
     String getSystemSerialNumber();
 
     /**