+++ oshi-json/src/test/java/oshi/json/SystemInfoTest.java
@@ -156,7 +156,6 @@ public class SystemInfoTest {
         System.out.println(" " + processor.getLogicalProcessorCount() + " logical CPU(s)");
 
         System.out.println("Identifier: " + processor.getIdentifier());
-        System.out.println("Serial Num: " + processor.getSystemSerialNumber());
     }
 
     private static void printMemory(GlobalMemory memory) {